refactor: align fixtures
