+++ packages/apps/reputation-oracle/server/src/modules/cron-job/fixtures/index.ts
@@ -1,6 +1,7 @@
 import { faker } from '@faker-js/faker';
-import { CronJobEntity } from './cron-job.entity';
-import { CronJobType } from './constants';
+
+import { CronJobEntity } from '../cron-job.entity';
+import { CronJobType } from '../constants';
 
 const cronJobTypes = Object.values(CronJobType);
