+++ packages/apps/reputation-oracle/server/src/modules/web3/fixtures/index.ts
@@ -2,9 +2,10 @@ import { faker } from '@faker-js/faker';
 import {
   Web3ConfigService,
   Web3Network,
-} from '../../config/web3-config.service';
-import { generateEthWallet } from '../../../test/fixtures/web3';
-import { supportedChainIdsByNetwork } from './web3.service';
+} from '../../../config/web3-config.service';
+import { generateEthWallet } from '../../../../test/fixtures/web3';
+
+import { supportedChainIdsByNetwork } from '../web3.service';
 
 const testWallet = generateEthWallet();
